a2a_financial_agent/host_adk/pyproject.toml
[project]
name = "a2a-financial-agent"
version = "0.1.0"
description = "A multi-agent system to coordinate business expense approvals."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.3.0",
    "a2a-sdk>=0.2.8",
    "nest-asyncio>=1.6.0",
    "uvicorn>=0.34.3",
    "python-dotenv",
    "click",
    "openai>=1.14.2",
    "httpx",
    "litellm>=1.72.4"
]
